// src/graphql/models/person.model.ts
export interface Person {
  name: string;
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  homeworld: string; // Planet
  films: string[]; // Film
  species: string[]; // Specie
  vehicles: string[]; // Vehicle
  starships: string[]; // Starship
  created: string;
  edited: string;
  url: string;
}